refactor(app): main を composition root に縮小
