+++ src/app/mod.rs
@@ -2,3 +2,5 @@
 
 pub(crate) mod diagnostics;
 pub(crate) mod diff_viewer;
+pub(crate) mod logging;
+pub(crate) mod terminal_mode;
